requirements.txt
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
coverage==5.5
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2.3
django-heroku==0.3.1
djangorestframework==3.12.4
generics==3.3.1
gunicorn==20.1.0
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
